Rename files
Renamed file structure and references from Maxar_OGC_SDK to Maxar_OGC
