+++ ogc-sdk/MANIFEST
@@ -1,0 +1,11 @@
+# file GENERATED by distutils, do NOT edit
+setup.cfg
+setup.py
+Maxar_OGC\__init__.py
+Maxar_OGC\auth.py
+Maxar_OGC\interface.py
+Maxar_OGC\process.py
+Maxar_OGC\wcs.py
+Maxar_OGC\wfs.py
+Maxar_OGC\wms.py
+Maxar_OGC\wmts.py
