+++ ogc-sdk/Maxar_OGC/auth.py
